@@ -1,14 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net471;netstandard2.0</TargetFrameworks>
-    <TargetFrameworks>netcoreapp2.0</TargetFrameworks>
+    <TargetFrameworks>netcoreapp2.2</TargetFrameworks>
     <AssemblyName>Lib</AssemblyName>
     <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
-
-    <DisableImplicitFSharpCoreReference>true</DisableImplicitFSharpCoreReference>
-    <DisableImplicitSystemValueTupleReference>true</DisableImplicitSystemValueTupleReference>
   </PropertyGroup>
   
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
@@ -25,11 +21,8 @@
   
   <ItemGroup>
     <Reference Include="netstandard" />
-    <Reference Condition="'$(TargetFramework)' == 'netstandard2.0'" Include="FSharp.Data.SqlClient">
-      <HintPath>..\..\..\bin\netstandard2.0\FSharp.Data.SqlClient.dll</HintPath>
-    </Reference>
-    <Reference Condition="'$(TargetFramework)' == 'net471'" Include="FSharp.Data.SqlClient">
-      <HintPath>..\..\..\bin\net40\FSharp.Data.SqlClient.dll</HintPath>
+    <Reference Condition="'$(TargetFramework)' == 'netstandard2.2'" Include="FSharp.Data.SqlClient">
+      <HintPath>..\..\..\bin\netstandard2.2\FSharp.Data.SqlClient.dll</HintPath>
     </Reference>
   </ItemGroup>
   
