@@ -2,19 +2,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
     <DisableImplicitFSharpCoreReference>true</DisableImplicitFSharpCoreReference>
-
-    <DisableImplicitFSharpCoreReference>true</DisableImplicitFSharpCoreReference>
-    <DisableImplicitSystemValueTupleReference>true</DisableImplicitSystemValueTupleReference>
   </PropertyGroup>
-  
   <ItemGroup>
     <Reference Include="FSharp.Data.SqlClient">
-      <HintPath>..\..\..\bin\netstandard2.0\FSharp.Data.SqlClient.dll</HintPath>
+      <HintPath>..\..\..\bin\netstandard2.2\FSharp.Data.SqlClient.dll</HintPath>
     </Reference>
   </ItemGroup>
-  
   <ItemGroup>
     <None Include="paket.references" />
     <Compile Include="Program.fs" />
@@ -22,8 +17,5 @@
   <ItemGroup>
     <None Include="paket.references" />
   </ItemGroup>
-  
-  <ItemGroup />
-  
   <Import Project="..\..\..\.paket\Paket.Restore.targets" />
 </Project>
